Move check_results to resnet18 file
